Update install phase folders

+++ nix/default.nix
@@ -83,9 +83,14 @@ in pkgs.gcc13Stdenv.mkDerivation rec {
   '';
 
   installPhase = ''
-    mkdir -p $out/lib $out/include
-    cp build/*storage* $out/lib/
-    cp library/libstorage.h $out/include/
+    if [ -f build/storage ]; then
+      mkdir -p $out/bin
+      cp build/storage $out/bin/
+    else
+      mkdir -p $out/lib $out/include
+      cp build/libstorage* $out/lib/
+      cp library/libstorage.h $out/include/
+    fi
   '';
 
   meta = with pkgs.lib; {
